requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
openai>=1.0.0